ISARA: more detailed emulation of the dewar lid
Emulate a more continouse process of opening and closing the lid.
When lid is beeing open/closed, it will traverse intermediate positions,
thus emulting the real lid that changes it's angle while being opened/closed.

Also, not it's possible to reverse the movement of the lid, event if it's
already moving.

Adds overlord hook, so it's now possible to monitor lid state on the
overlord port.