MD3 exporter: emulate 'WRTE' message
Emulate writing MD3 attributes with 'WRTE' message.

On WRTE message:

 - update internal value of  the attribute
 - emit event message about the changed value

If writing to a attribute of float type, treat it as a motor move. Motor
is moved in increments to the new value. The motor's 'State' attribute
is changed to 'Moving' and back to 'Ready' when we have reached our target
position.